Bumped dependencies; Refactor to be more like the other micro services;

+++ repositories/entities/file-entity.mjs
@@ -1,10 +1,10 @@
-import {numeric, sqliteTable, text} from 'drizzle-orm/sqlite-core';
+import {integer, sqliteTable, text} from 'drizzle-orm/sqlite-core';
 
 
 export const FileEntity = sqliteTable('files', {
 	id: text('id').notNull().unique().primaryKey(),
 	isDeleted: text('is_deleted'),
-	creationTs: numeric('creation_ts').notNull(),
-	lastModifiedTs: numeric('last_modified_ts').notNull(),
-	deletedTs: numeric('deleted_ts').notNull(),
+	creationTs: integer('creation_ts').notNull(),
+	lastModifiedTs: integer('last_modified_ts').notNull(),
+	deletedTs: integer('deleted_ts').notNull(),
 });
